Fixed bugs, Added logic for delete before upload - Periodical Strategy

+++ UploadService/Configurations/UploadStrategies/Implementations/PeriodicalStrategy.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Resources;
+using System.Threading.Tasks;
 using System.Timers;
 using UploadService.Configurations.UploadTypeConfgurations;
 using UploadService.Configurations.UploadTypeConfgurations.Implementations;
@@ -41,27 +42,62 @@ namespace UploadService.Configurations.UploadStrategies.Implementations
         private void OnTimedEvent(object sender, ElapsedEventArgs e)
         {
             
+           
             foreach (var folder in _foldersToUpload.Select(f => f.LocalFolderPath))
             {
+           
                 var remoteFolder = _foldersToUpload.Where(rf => rf.LocalFolderPath == folder).Select(rm => rm.RemoteFolder).FirstOrDefault();
                 var fileMask = _foldersToUpload.Where(rf => rf.LocalFolderPath == folder).Select(r => r.FileMask).FirstOrDefault();
-               // TODO Add other parameters
-                foreach (string filePath in Directory.GetFiles(folder))
+                
+                DirectoryInfo startDir = new DirectoryInfo(folder);
+                TraverseDirectory(startDir,fileMask, remoteFolder);
+              
+
+            }
+        }
+        
+        public void TraverseDirectory(DirectoryInfo directoryInfo, string fileMask, string remoteFolder)
+        {
+            //var subdirectories = directoryInfo.EnumerateDirectories();
+
+            /*foreach (var subdirectory in subdirectories)
+            {
+                TraverseDirectory(subdirectory, fileMask,remoteFolder);
+            }*/
+
+            var files = directoryInfo.EnumerateFiles();
+          
+            foreach (var file in files)
+            {
+                HandleFile(file,fileMask, remoteFolder, directoryInfo.ToString());
+            }
+        }
+
+        void HandleFile(FileInfo file,string fileMask, string remoteFolder, string directory)
+        {
+            
+           
+            string[] arraytmp  = file.ToString().Split('/');
+            var fileName = arraytmp[arraytmp.Length - 1];
+            //TODO check before
+            if (fileName.Contains(fileMask))
+            {
+                
+                if (_client.checkIfFileExists($"{"home/katarina/" + remoteFolder + "/"}{fileName}"))
                 {
-                    string[] arraytmp  = filePath.Split('/');
-                    var fileName = arraytmp[arraytmp.Length - 1];
-                    if (fileName.Contains(fileMask))
-                    {
-                        //TODO Ask async
-                        _client.delete($"{"home/katarina/" + remoteFolder + "/"}{fileName}");
-                        _client.UploadFile($"{"home/katarina/" + remoteFolder + "/"}{fileName}", $"{folder + "/"}{fileName}");
-                    }
-             
-                   
+                    _client.delete($"{"home/katarina/" + remoteFolder + "/"}{fileName}");
+                    _client.UploadFile($"{"home/katarina/" + remoteFolder + "/"}{fileName}",
+                        $"{directory.ToString() + "/"}{fileName}");
+                }
+                else 
+                {
+                    _client.UploadFile($"{"home/katarina/" + remoteFolder + "/"}{fileName}", $"{directory.ToString() + "/"}{fileName}");
                 }
 
-
+                
             }
         }
     }
 }
+
+
